PLIC working and ready to use
