+++ samples/hello_world/prj.conf
@@ -3,13 +3,7 @@
 CONFIG_I2C_SECURE_IOT=y
 CONFIG_MULTITHREADING=y
 CONFIG_GPIO=y
-CONFIG_GPIO_SHAKTI=y
-CONFIG_PLIC=y
 CONFIG_SERIAL=y
-CONFIG_SHAKTI_PLIC=y
-
-CONFIG_GEN_ISR_TABLES=y
-CONFIG_GEN_SW_ISR_TABLE=y
 
 # CONFIG_MAX_IRQ_PER_AGGREGATOR=58
